refactor index page
